fix numpy misc functions
