+++ brainunit/math/_numpy_linear_algebra.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 from __future__ import annotations
 
-from typing import (Union, Any, Optional)
+from typing import (Union, Any, Optional, Tuple)
 
 import jax
 import jax.numpy as jnp
@@ -27,7 +27,7 @@ from .._misc import set_module_as
 
 __all__ = [
   # linear algebra
-  'dot', 'vdot', 'inner', 'outer', 'kron', 'matmul', 'trace',
+  'dot', 'vdot', 'inner', 'outer', 'kron', 'matmul', 'tensordot', 'trace',
 
 ]
 
@@ -254,6 +254,61 @@ def matmul(
                                  precision=precision,
                                  preferred_element_type=preferred_element_type)
 
+
+@set_module_as('brainunit.math')
+def tensordot(
+    a: Union[jax.typing.ArrayLike, Quantity],
+    b: Union[jax.typing.ArrayLike, Quantity],
+    axes: Union[int, Tuple[int, int]] = 2,
+    precision: Any = None,
+    preferred_element_type: Optional[jax.typing.DTypeLike] = None
+) -> Union[jax.Array, Quantity]:
+  """
+  Compute tensor dot product along specified axes.
+
+  Given two tensors, `a` and `b`, and an array_like object containing
+  two array_like objects, ``(a_axes, b_axes)``, sum the products of
+  `a`'s and `b`'s elements (components) over the axes specified by
+  ``a_axes`` and ``b_axes``. The third argument can be a single non-negative
+  integer_like scalar, ``N``; if it is such, then the last ``N`` dimensions
+  of `a` and the first ``N`` dimensions of `b` are summed over.
+
+  Parameters
+  ----------
+  a, b : array_like, Quantity
+    Tensors to "dot".
+
+  axes : int or (2,) array_like
+    * integer_like
+      If an int N, sum over the last N axes of `a` and the first N axes
+      of `b` in order. The sizes of the corresponding axes must match.
+    * (2,) array_like
+      Or, a list of axes to be summed over, first sequence applying to `a`,
+      second to `b`. Both elements array_like must be of the same length.
+  precision : Optional. Either ``None``, which means the default precision for
+    the backend, a :class:`~jax.lax.Precision` enum value
+    (``Precision.DEFAULT``, ``Precision.HIGH`` or ``Precision.HIGHEST``), a
+    string (e.g. 'highest' or 'fastest', see the
+    ``jax.default_matmul_precision`` context manager), or a tuple of two
+    :class:`~jax.lax.Precision` enums or strings indicating precision of
+    ``lhs`` and ``rhs``.
+  preferred_element_type : Optional. Either ``None``, which means the default
+    accumulation type for the input types, or a datatype, indicating to
+    accumulate results to and return a result with that datatype.
+
+  Returns
+  -------
+  output : ndarray, Quantity
+    The tensor dot product of the input.
+
+    This is a quantity if the product of the units of `a` and `b` is not dimensionless.
+  """
+  return _fun_change_unit_binary(jnp.tensordot,
+                                 lambda x, y: x * y,
+                                 a, b,
+                                 axes=axes,
+                                 precision=precision,
+                                 preferred_element_type=preferred_element_type)
 
 @set_module_as('brainunit.math')
 def trace(
