+++ brainunit/math/_numpy_misc.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 
 import collections
 from collections.abc import Sequence
-from typing import (Callable, Union, Tuple, Any, Optional)
+from typing import (Callable, Union, Optional)
 
 import jax
 import jax.numpy as jnp
@@ -25,8 +25,6 @@ import numpy as np
 import opt_einsum
 from jax import Array
 
-from ._numpy_change_unit import _fun_change_unit_binary
-from ._numpy_keep_unit import _fun_keep_unit_unary
 from .._base import (DIMENSIONLESS,
                      Quantity,
                      fail_for_dimension_mismatch,
@@ -35,7 +33,6 @@ from .._base import (DIMENSIONLESS,
 from .._misc import set_module_as
 
 __all__ = [
-
   # constants
   'e', 'pi', 'inf',
 
@@ -43,9 +40,7 @@ __all__ = [
   'dtype', 'finfo', 'iinfo',
 
   # more
-  'broadcast_arrays', 'broadcast_shapes',
-  'einsum', 'gradient', 'intersect1d', 'nan_to_num', 'nanargmax', 'nanargmin',
-  'rot90', 'tensordot', 'frexp',
+  'broadcast_arrays', 'broadcast_shapes', 'einsum', 'gradient',
 ]
 
 # constants
@@ -302,304 +297,3 @@ def gradient(
     varargs = [v.value if isinstance(v, Quantity) else v for v in varargs]
     result_list = jnp.gradient(f, *varargs, axis=axis)
     return [Quantity(r, dim=unit) if unit is not None else r for r, unit in zip(result_list, unit_list)]
-
-
-@set_module_as('brainunit.math')
-def intersect1d(
-    ar1: Union[jax.typing.ArrayLike, Quantity],
-    ar2: Union[jax.typing.ArrayLike, Quantity],
-    assume_unique: bool = False,
-    return_indices: bool = False
-) -> Union[jax.Array, Quantity, tuple[Union[jax.Array, Quantity], jax.Array, jax.Array]]:
-  """
-  Find the intersection of two arrays.
-
-  Return the sorted, unique values that are in both of the input arrays.
-
-  Parameters
-  ----------
-  ar1, ar2 : array_like, Quantity
-    Input arrays. Will be flattened if not already 1D.
-  assume_unique : bool
-    If True, the input arrays are both assumed to be unique, which
-    can speed up the calculation.  If True but ``ar1`` or ``ar2`` are not
-    unique, incorrect results and out-of-bounds indices could result.
-    Default is False.
-  return_indices : bool
-    If True, the indices which correspond to the intersection of the two
-    arrays are returned. The first instance of a value is used if there are
-    multiple. Default is False.
-
-  Returns
-  -------
-  intersect1d : ndarray, Quantity
-    Sorted 1D array of common and unique elements.
-  comm1 : ndarray
-    The indices of the first occurrences of the common values in `ar1`.
-    Only provided if `return_indices` is True.
-  comm2 : ndarray
-    The indices of the first occurrences of the common values in `ar2`.
-    Only provided if `return_indices` is True.
-  """
-  fail_for_dimension_mismatch(ar1, ar2, 'intersect1d')
-  unit = None
-  if isinstance(ar1, Quantity):
-    unit = ar1.dim
-  ar1 = ar1.value if isinstance(ar1, Quantity) else ar1
-  ar2 = ar2.value if isinstance(ar2, Quantity) else ar2
-  result = jnp.intersect1d(ar1, ar2, assume_unique=assume_unique, return_indices=return_indices)
-  if return_indices:
-    if unit is not None:
-      return Quantity(result[0], dim=unit), result[1], result[2]
-    else:
-      return result
-  else:
-    if unit is not None:
-      return Quantity(result, dim=unit)
-    else:
-      return result
-
-
-@set_module_as('brainunit.math')
-def nan_to_num(
-    x: Union[jax.typing.ArrayLike, Quantity],
-    nan: float | Quantity = None,
-    posinf: float | Quantity = None,
-    neginf: float | Quantity = None
-) -> Union[jax.Array, Quantity]:
-  """
-  Replace NaN with zero and infinity with large finite numbers (default
-  behaviour) or with the numbers defined by the user using the `nan`,
-  `posinf` and/or `neginf` keywords.
-
-  If `x` is inexact, NaN is replaced by zero or by the user defined value in
-  `nan` keyword, infinity is replaced by the largest finite floating point
-  values representable by ``x.dtype`` or by the user defined value in
-  `posinf` keyword and -infinity is replaced by the most negative finite
-  floating point values representable by ``x.dtype`` or by the user defined
-  value in `neginf` keyword.
-
-  For complex dtypes, the above is applied to each of the real and
-  imaginary components of `x` separately.
-
-  If `x` is not inexact, then no replacements are made.
-
-  Parameters
-  ----------
-  x : scalar, array_like or Quantity
-    Input data.
-  nan : int, float, optional
-    Value to be used to fill NaN values. If no value is passed
-    then NaN values will be replaced with 0.0.
-  posinf : int, float, optional
-    Value to be used to fill positive infinity values. If no value is
-    passed then positive infinity values will be replaced with a very
-    large number.
-  neginf : int, float, optional
-    Value to be used to fill negative infinity values. If no value is
-    passed then negative infinity values will be replaced with a very
-    small (or negative) number.
-
-  Returns
-  -------
-  out : ndarray, Quantity
-    `x`, with the non-finite values replaced. If `copy` is False, this may
-    be `x` itself.
-  """
-  if isinstance(x, Quantity):
-    nan = Quantity(0.0, dim=x.dim) if nan is None else nan
-    posinf = Quantity(jnp.finfo(x.dtype).max, dim=x.dim) if posinf is None else posinf
-    neginf = Quantity(jnp.finfo(x.dtype).min, dim=x.dim) if neginf is None else neginf
-    return Quantity(jnp.nan_to_num(x.value, nan=nan.value, posinf=posinf.value, neginf=neginf.value), dim=x.dim)
-  else:
-    nan = 0.0 if nan is None else nan
-    posinf = jnp.inf if posinf is None else posinf
-    neginf = -jnp.inf if neginf is None else neginf
-    return jnp.nan_to_num(x, nan=nan, posinf=posinf, neginf=neginf)
-
-
-@set_module_as('brainunit.math')
-def rot90(
-    m: Union[jax.typing.ArrayLike, Quantity],
-    k: int = 1,
-    axes: Tuple[int, int] = (0, 1)
-) -> Union[
-  jax.Array, Quantity]:
-  """
-  Rotate an array by 90 degrees in the plane specified by axes.
-
-  Rotation direction is from the first towards the second axis.
-
-  Parameters
-  ----------
-  m : array_like, Quantity
-    Array of two or more dimensions.
-  k : integer
-    Number of times the array is rotated by 90 degrees.
-  axes : (2,) array_like
-    The array is rotated in the plane defined by the axes.
-    Axes must be different.
-
-  Returns
-  -------
-  y : ndarray, Quantity
-    A rotated view of `m`.
-
-    This is a quantity if `m` is a quantity.
-  """
-  return _fun_keep_unit_unary(jnp.rot90, m, k=k, axes=axes)
-
-
-@set_module_as('brainunit.math')
-def tensordot(
-    a: Union[jax.typing.ArrayLike, Quantity],
-    b: Union[jax.typing.ArrayLike, Quantity],
-    axes: Union[int, Tuple[int, int]] = 2,
-    precision: Any = None,
-    preferred_element_type: Optional[jax.typing.DTypeLike] = None
-) -> Union[jax.Array, Quantity]:
-  """
-  Compute tensor dot product along specified axes.
-
-  Given two tensors, `a` and `b`, and an array_like object containing
-  two array_like objects, ``(a_axes, b_axes)``, sum the products of
-  `a`'s and `b`'s elements (components) over the axes specified by
-  ``a_axes`` and ``b_axes``. The third argument can be a single non-negative
-  integer_like scalar, ``N``; if it is such, then the last ``N`` dimensions
-  of `a` and the first ``N`` dimensions of `b` are summed over.
-
-  Parameters
-  ----------
-  a, b : array_like, Quantity
-    Tensors to "dot".
-
-  axes : int or (2,) array_like
-    * integer_like
-      If an int N, sum over the last N axes of `a` and the first N axes
-      of `b` in order. The sizes of the corresponding axes must match.
-    * (2,) array_like
-      Or, a list of axes to be summed over, first sequence applying to `a`,
-      second to `b`. Both elements array_like must be of the same length.
-  precision : Optional. Either ``None``, which means the default precision for
-    the backend, a :class:`~jax.lax.Precision` enum value
-    (``Precision.DEFAULT``, ``Precision.HIGH`` or ``Precision.HIGHEST``), a
-    string (e.g. 'highest' or 'fastest', see the
-    ``jax.default_matmul_precision`` context manager), or a tuple of two
-    :class:`~jax.lax.Precision` enums or strings indicating precision of
-    ``lhs`` and ``rhs``.
-  preferred_element_type : Optional. Either ``None``, which means the default
-    accumulation type for the input types, or a datatype, indicating to
-    accumulate results to and return a result with that datatype.
-
-  Returns
-  -------
-  output : ndarray, Quantity
-    The tensor dot product of the input.
-
-    This is a quantity if the product of the units of `a` and `b` is not dimensionless.
-  """
-  return _fun_change_unit_binary(jnp.tensordot,
-                                 lambda x, y: x * y,
-                                 a, b,
-                                 axes=axes,
-                                 precision=precision,
-                                 preferred_element_type=preferred_element_type)
-
-
-@set_module_as('brainunit.math')
-def nanargmax(
-    a: Union[jax.typing.ArrayLike, Quantity],
-    axis: int = None,
-    keepdims: bool = False
-) -> jax.Array:
-  """
-  Return the indices of the maximum values in the specified axis ignoring
-  NaNs. For all-NaN slices ``ValueError`` is raised. Warning: the
-  results cannot be trusted if a slice contains only NaNs and -Infs.
-
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  axis : int, optional
-    Axis along which to operate.  By default flattened input is used.
-  keepdims : bool, optional
-    If this is set to True, the axes which are reduced are left
-    in the result as dimensions with size one. With this option,
-    the result will broadcast correctly against the array.
-
-  Returns
-  -------
-  index_array : ndarray
-    An array of indices or a single index value.
-  """
-  return _fun_require_same_unit(jnp.nanargmax,
-                                a,
-                                return_quantity=False,
-                                axis=axis,
-                                keepdims=keepdims)
-
-
-@set_module_as('brainunit.math')
-def nanargmin(
-    a: Union[jax.typing.ArrayLike, Quantity],
-    axis: int = None,
-    keepdims: bool = False
-) -> jax.Array:
-  """
-  Return the indices of the minimum values in the specified axis ignoring
-  NaNs. For all-NaN slices ``ValueError`` is raised. Warning: the results
-  cannot be trusted if a slice contains only NaNs and Infs.
-
-  Parameters
-  ----------
-  a : array_like, Quantity
-    Input data.
-  axis : int, optional
-    Axis along which to operate.  By default flattened input is used.
-  keepdims : bool, optional
-    If this is set to True, the axes which are reduced are left
-    in the result as dimensions with size one. With this option,
-    the result will broadcast correctly against the array.
-
-  Returns
-  -------
-  index_array : ndarray
-    An array of indices or a single index value.
-  """
-  return _fun_require_same_unit(jnp.nanargmin,
-                                a,
-                                return_quantity=False,
-                                axis=axis,
-                                keepdims=keepdims)
-
-
-@set_module_as('brainunit.math')
-def frexp(
-    x: Union[Quantity, jax.typing.ArrayLike]
-) -> Tuple[jax.Array, jax.Array]:
-  """
-  Decompose the elements of x into mantissa and twos exponent.
-
-  Returns (`mantissa`, `exponent`), where ``x = mantissa * 2**exponent``.
-  The mantissa lies in the open interval(-1, 1), while the twos
-  exponent is a signed integer.
-
-  Parameters
-  ----------
-  x : array_like, Quantity
-    Array of numbers to be decomposed.
-
-  Returns
-  -------
-  mantissa : ndarray
-    Floating values between -1 and 1.
-    This is a scalar if `x` is a scalar.
-  exponent : ndarray
-    Integer exponents of 2.
-    This is a scalar if `x` is a scalar.
-  """
-  assert not isinstance(x, Quantity) or is_unitless(x), "Input must be unitless"
-  x = x.value if isinstance(x, Quantity) else x
-  return jnp.frexp(x)
